tracing package and handling nil messages
